AccountController Api Eklendi

+++ BankAppCoreWebApi/BankAppCoreWebApi/Models/Account.cs
@@ -5,16 +5,16 @@ using System.Threading.Tasks;
 
 namespace BankAppCoreWebApi.Models
 {
-	public class Account:BaseEntity
+	public class Account : BaseEntity
 	{
-		public  int Id { get; set; }
-		public int  customerId{ get; set; }
-		public int balance { get; set; }
-		public int blockageAmount { get; set; }
-		public int netBalance { get; set; }
+		public int Id { get; set; }
+		public int customerId { get; set; }
+		public decimal balance { get; set; }
+		public decimal blockageAmount { get; set; }
+		public decimal netBalance { get; set; }
 		public DateTime openingDate { get; set; }
 		public DateTime lastTransactionDate { get; set; }
-		public bool status{ get; set; }
+		public bool status { get; set; }
 	}
 
 }
